type conversion

+++ db/TagTable.sql
@@ -1,7 +1,7 @@
 create table TagTable(
-    tagId int not null auto_increment,
+    tagId bigint not null auto_increment,
     tagName char(20) not null,
-    categoryId int not null,
+    categoryId bigint not null,
     primary key(tagId),
     foreign key (categoryId) references CategoryTable(categoryId)
 );
